Add vision unitTests template for windows
Outline for loading opencv dll to allow vision unit tests 
to be run in windows.
Could easily be expanded to support mac.

If dll is not detected, android studio gives links 
to download the windows dll.

Should detect mac vs pc when loading the openCV library.

Note: There are no tests for the actual pipelines and 
detectors, only for the opencv library itself.