AngleHighGoalPipeline test template
"Template" because there are no assertions, and no 
tests that can fail. However, the pipeline is run, and the 
outputs are saved.

Note that for all of these tests with an array of test images, 
the problem is that the test will fail on the first image 
failure, and not run the remaining images through the 
pipeline.

A more appropriate solution might be to use 
parameterized tests.


https://github.com/junit-team/junit4/wiki/parameterized-tests

https://junit.org/junit5/docs/current/user-guide/#writing-tests-parameterized-tests